extract type definitions to an external file
