+++ integration/types.go
@@ -1,0 +1,16 @@
+package integration
+
+// MobileClientSpec represents a mobile client application
+type MobileClientSpec struct {
+	ID            string
+	Name          string
+	APIKey        string
+	ClientType    string
+	AppIdentifier string
+	Namespace     string
+}
+
+// MobileClientJSON represents a mobile client application
+type MobileClientJSON struct {
+	Spec MobileClientSpec
+}
